+++ src/lib/BlackHole.ts
@@ -1,11 +1,11 @@
-import { Writable, WritableOptions } from "stream"
+import { Writable, WritableOptions } from 'stream'
 
 export class BlackHole extends Writable {
     constructor(options?: WritableOptions) {
-        super(options);
+        super(options)
     }
 
     _write(_chunk: any, _encoding: BufferEncoding, callback: (error?: Error | null) => void): void {
-        callback(); // eat the input
+        callback() // eat the input
     }
 }