+++ src/lib/proto/http.ts
@@ -3,44 +3,45 @@ import * as request from 'request'
 import isRelativeUrl = require('is-relative-url')
 import ms = require('ms')
 
-import { Callback, Options, Protocol, staticImplements } from "../types"
+import { Callback, Options, Protocol, staticImplements } from '../types'
 import { LinkCheckResult } from '../LinkCheckResult'
 import { BlackHole } from '../BlackHole'
 
 @staticImplements<Protocol>()
 export class HttpProtocol {
-
     public static check(link: string, opts: Options, callback: Callback): void {
-
         // default request timeout set to 10s if not provided in options
-        const timeout = opts.timeout || '10s';
+        const timeout = opts.timeout || '10s'
 
         const requestOptions: request.OptionsWithUri = {
             uri: encodeURI(link),
             headers: {
                 // override 'User-Agent' (some sites return `401` when the user-agent isn't a web browser)
-                'User-Agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_11_1) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/47.0.2526.106 Safari/537.36'
+                'User-Agent':
+                    'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_11_1) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/47.0.2526.106 Safari/537.36',
             },
             maxRedirects: 8,
             strictSSL: false,
             timeout: ms(timeout) as number,
-        };
+        }
 
         if (opts.baseUrl && isRelativeUrl(link)) {
-            requestOptions.baseUrl = opts.baseUrl;
+            requestOptions.baseUrl = opts.baseUrl
         }
 
         if (opts.headers) {
-            Object.assign(requestOptions.headers, opts.headers);
+            Object.assign(requestOptions.headers, opts.headers)
         }
 
         HttpProtocol.doCheckWithRetry(link, opts, callback, requestOptions)
     }
 
+    // prettier-ignore
     private static doCheckWithRetry(link: string, opts: Options, callback: Callback, requestOptions: request.OptionsWithUri, attempts: number = 0, additionalMessage?: string): void {
         HttpProtocol.doHeadWithRetry(link, opts, callback, requestOptions, attempts, additionalMessage)
     }
 
+    // prettier-ignore
     public static doHeadWithRetry(link: string, opts: Options, callback: Callback, requestOptions: request.OptionsWithUri, attempts: number, additionalMessage?: string) {
         request.head(requestOptions, (err: any, res: request.Response, body: any): void => {
 
@@ -56,6 +57,7 @@ export class HttpProtocol {
         })
     }
 
+    // prettier-ignore
     public static doGetWithRetry(link: string, opts: Options, callback: Callback, requestOptions: request.OptionsWithUri, attempts: number, additionalMessage?: string) {
         // retry on 429 http code flag is false by default if not provided in options
         const retryOn429 = opts.retryOn429 || false;
@@ -82,6 +84,7 @@ export class HttpProtocol {
                     retryInMs = fallbackRetryDelayInMs;
                 }
                 // Recurse back after the retry timeout has elapsed (incrementing our attempts to avoid an infinite loop)
+                // prettier-ignore
                 setTimeout(HttpProtocol.doCheckWithRetry, retryInMs, link, opts, callback, requestOptions, attempts + 1, additionalMessage);
             } else {
                 const linkErr = 
@@ -112,30 +115,30 @@ export class HttpProtocol {
             // tcort/link-check will then stop the support of non standard 'retry-after' header for releases greater or equal to 4.7.0.
             // all this 'isNaN' part and the additionalMessage will then be removed from the code.
             additionalMessage =
-                "Server returned a non standard \'retry-after\' header. "
-                + "Non standard \'retry-after\' header will not work after link-check 4.7.0 release. "
-                + "See https://github.com/tcort/link-check/releases/tag/v4.5.2 release note for details.";
+                "Server returned a non standard 'retry-after' header. " +
+                "Non standard 'retry-after' header will not work after link-check 4.7.0 release. " +
+                'See https://github.com/tcort/link-check/releases/tag/v4.5.2 release note for details.'
 
-            let buf = '';
-            let letter = false;
+            let buf = ''
+            let letter = false
             for (const c of retryAfterStr) {
                 if (isNaN(Number(c))) {
-                    letter = true;
-                    buf += c;
+                    letter = true
+                    buf += c
                 } else {
                     if (letter) {
                         retryInMs += ms(buf.trim()) as number
-                        buf = '';
+                        buf = ''
                     }
-                    letter = false;
-                    buf += c;
+                    letter = false
+                    buf += c
                 }
             }
-            retryInMs = ms(buf.trim());
+            retryInMs = ms(buf.trim())
         } else {
             // standard compliant header value conversion to milliseconds
-            const secondsToMilisecondsMultiplier = 1000;
-            retryInMs = parseFloat(retryAfterStr) * secondsToMilisecondsMultiplier;
+            const secondsToMilisecondsMultiplier = 1000
+            retryInMs = parseFloat(retryAfterStr) * secondsToMilisecondsMultiplier
         }
 
         return {
@@ -144,4 +147,3 @@ export class HttpProtocol {
         }
     }
 }
-